# Faster MutatingScope->mergeWith(Scope)