profilled the `src/PhpSpreadsheet/Reader/Xlsx.php` case further:

we spent most of the time with merging scope objects
profile can be seen here: https://blackfire.io/profiles/0293ba31-80fc-403f-9452-bfda15b7c7c7/graph

before this PR
```
➜  PhpSpreadsheet git:(master) ✗ hyperfine 'php ../phpstan-src/bin/phpstan analyze src/PhpSpreadsheet/Reader/Xlsx.php --debug'
Benchmark 1: php ../phpstan-src/bin/phpstan analyze src/PhpSpreadsheet/Reader/Xlsx.php --debug
  Time (mean ± σ):      8.659 s ±  0.029 s    [User: 8.479 s, System: 0.175 s]
  Range (min … max):    8.611 s …  8.701 s    10 runs
```

after this PR
```
➜  PhpSpreadsheet git:(master) ✗ hyperfine 'php ../phpstan-src/bin/phpstan analyze src/PhpSpreadsheet/Reader/Xlsx.php --debug'
Benchmark 1: php ../phpstan-src/bin/phpstan analyze src/PhpSpreadsheet/Reader/Xlsx.php --debug
  Time (mean ± σ):      7.462 s ±  0.073 s    [User: 7.294 s, System: 0.164 s]
  Range (min … max):    7.349 s …  7.550 s    10 runs
```